+++ docs/changelog.rst
@@ -1,9 +1,23 @@
 Changelog
 ++++++++++
 
+.. **0.2.0 (08/2024)**
+
+.. Addressing JOSS review comments:
+
+.. - placeholder
+
+.. Other changes:
+
+.. - The `log_jitter` term now represents the *standard deviation* of the added noise, not the variance. This brings it in line with the definition of measurement uncertainties.
+.. - When dealing with a projected 2D ellipse (i.e., not a tidally locked planet) ``projected_r`` has been replaced with ``projected_effective_r``. Previously there was a strong degeneracy between ``projected_r`` and ``projected_f``, since the combination of the two dictated the area of the ellipse, and therefore transit depth, which fits are much more sensitive to than to slight deviations to ingress/egress shape. ``projected_effective_r`` is the radius of a circle with the same area as the ellipse, and is therefore a more physically meaningful parameter to fit for. 
+.. - Added fit_limb_darkening_profile() as a convenience function for approximating limb darkening profiles computed via stellar grids as high-order polynomial laws that `squishyplanet` can use.
+
+
 **0.1.2 (05/2024)**
 
-Fixed issue #7. Previously, calls to ``OblateSystem.lightcurve()`` after initializing the ``OblateSystem`` object were not recalculating true anomalies even if the ``period`` and/or ``t_peri`` changed.
+Fixed issue `#7
+<https://github.com/ben-cassese/squishyplanet/issues/7/>`_. Previously, calls to ``OblateSystem.lightcurve()`` after initializing the ``OblateSystem`` object were not recalculating true anomalies even if the ``period`` and/or ``t_peri`` changed.
 
 **0.1.1 (05/2024)**